started programming button to accept challenge and set up backend for new game

+++ app/channels/start_challenge_channel.rb
@@ -3,10 +3,16 @@ class StartChallengeChannel < ApplicationCable::Channel
     stream_from "#{params[:user_id]}_challenge"
   end
 
+  def receive(data)
+    ActionCable.server.broadcast("#{params[:user_id]}_challenge", data)
+  end
+
   def unsubscribed
     user = User.find(current_user.id)
-    id = user.challenge.id
-    user.challenge.destroy
+    if user.challenge
+      id = user.challenge.id
+      user.challenge.destroy
+    end
     ActionCable.server.broadcast("challenge_channel", { challenge: { deleted: true, id: id} })
     stop_stream_from "#{params[:user_id]}_challenge"
   end
